move child class to separate file
