+++ src/backtesting/backtester.py
@@ -79,40 +79,5 @@ class Backtester:
         return self._summary.get_data()
 
 
-class DailyMarketReplay(Backtester):
-    def run_back_testing(self):
-        self._has_tested = True
-        date = self._start_date
-        while date < self._end_date:
-            self._update_by_symbol(self._self_agent, date)
-            for agent in self._agents:
-                self._update_by_symbol(agent, date)
-            date += timedelta(days=1)
-
-    def _update_by_symbol(self, agent: Agent, date: datetime):
-        for symbol in agent.get_symbols():
-            if not agent.need_update_for(date, symbol):
-                continue
-            da_result = retrieve_stock(symbol, date)
-            if da_result.is_successful:
-                stock: Stock
-                stock = da_result.data
-                # if np.isnan(stock.current_price):
-                #     continue
-                new_data = (stock.get_price().price(), self._simulate_premiums(stock, date))
-                self._self_agent.update(symbol, new_data, date)
-
-    @staticmethod
-    def _simulate_premiums(stock: Stock, date: datetime, is_call=False):
-        strikes = simulate_strikes(stock.get_price().price())
-        result = dict[float, float]()
-        for strike in strikes:
-            expiry_date = date + timedelta(days=30)
-            rf = retrieve_rf(expiry_date).data
-            premium = bsm_pricing(stock, strike, expiry_date, [], rf, is_call)
-            result[strike] = round(premium, 2)
-        return result
-
-
 class MultiAgent(Backtester):
     pass
